@@ -3,6 +3,7 @@ package com.keepu.webAPI.controller;
 import com.keepu.webAPI.dto.request.CreateAuthCodeRequest;
 import com.keepu.webAPI.dto.response.AuthCodeResponse;
 import com.keepu.webAPI.model.AuthCode;
+import com.keepu.webAPI.model.enums.AuthCodeType;
 import com.keepu.webAPI.service.AuthCodeService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
@@ -23,12 +24,20 @@ public class AuthCodeController {
     public ResponseEntity<AuthCodeResponse> create(@Valid @RequestBody CreateAuthCodeRequest request) {
         return ResponseEntity.ok(authCodeService.createAuthCode(request));
     }
-    @PutMapping("/{code}")
-    public void update(@PathVariable String code) {
-        authCodeService.updateAuthCode(code);
+    @PatchMapping
+    public ResponseEntity<AuthCodeResponse> update(@RequestParam String code) {
+        return ResponseEntity.ok (authCodeService.updateAuthCode(code));
     }
     @GetMapping
     public ResponseEntity<List<AuthCode>> getAll() {
         return ResponseEntity.ok(authCodeService.getAllAuthCodes());
+    }
+    @GetMapping("/user/{id}")
+    public ResponseEntity<AuthCode> getByUserId(@PathVariable Long id, @RequestParam AuthCodeType codeType) {
+        return ResponseEntity.ok(authCodeService.getAuthCodeById(id, codeType));
+    }
+    @GetMapping("/{code}")
+    public ResponseEntity<AuthCode> getByCode(@PathVariable String code) {
+        return ResponseEntity.ok(authCodeService.getAuthCodeByCode(code));
     }
 }
