@@ -4,7 +4,7 @@ import jakarta.validation.constraints.NotNull;
 
 public record CreateInvitationCodeRequest(
         @NotNull(message = "El ID del usuario no puede ser nulo")
-        Integer userId,
+        Long userId,
 
         @NotNull(message="El nombre del hijo no puede se null")
         String childName,
