@@ -1,7 +1,6 @@
 package com.keepu.webAPI.dto.response;
 
 public record ChildrenResponse(
-        Integer id,
-        Integer userId,
+        Long id,
         Integer age
 ) {}
