@@ -3,7 +3,7 @@ package com.keepu.webAPI.dto.response;
 import com.keepu.webAPI.model.enums.RelationshipType;
 
 public record ParentChildrenResponse(
-        Integer parentId,
-        Integer childId,
+        Long parentId,
+        Long childId,
         RelationshipType relationshipType
 ) {}
