@@ -1,7 +1,6 @@
 package com.keepu.webAPI.dto.response;
 
 public record ParentResponse(
-        Integer id,
-        Integer phone,
-        Integer userId
+        Long id,
+        Integer phone
 ) {}
