@@ -2,7 +2,7 @@ package com.keepu.webAPI.dto.response;
 
 public record UserCourseProgressResponse(
         Integer id,
-        Integer userId,
+        Long userId,
         Integer courseId,
         Double progressPercentage,
         Boolean isCompleted
