@@ -2,6 +2,7 @@ package com.keepu.webAPI.mapper;
 
 import com.keepu.webAPI.dto.request.CreateAuthCodeRequest;
 import com.keepu.webAPI.dto.response.AuthCodeResponse;
+import com.keepu.webAPI.model.UserAuth;
 import com.keepu.webAPI.model.enums.AuthCodeType;
 import com.keepu.webAPI.model.AuthCode;
 import com.keepu.webAPI.model.User;
@@ -16,7 +17,7 @@ public class AuthCodeMapper {
         }
         return new AuthCodeResponse(
                 authCode.getId(),
-                authCode.getUser().getId(),
+                authCode.getUserAuth().getUser() .getId(),
                 authCode.getCode(),
                 authCode.getCodeType(),
                 authCode.isUsed(),
@@ -24,13 +25,13 @@ public class AuthCodeMapper {
         );
     }
 
-    public AuthCode toAuthCodeEntity(CreateAuthCodeRequest request, User user) {
+    public AuthCode toAuthCodeEntity(CreateAuthCodeRequest request, UserAuth user) {
         if (request == null || user == null) {
             return null;
         }
 
         AuthCode authCode = new AuthCode();
-        authCode.setUser(user);
+        authCode.setUserAuth(user);
         authCode.setCodeType(request.authCodeType());
         return authCode;
     }
